refactor: extract service interfaces
