+++ src/application/interfaces/chat/ChatConfigService.errors.ts
@@ -1,0 +1,2 @@
+export class InvalidHistoryLimitError extends Error {}
+export class InvalidInterestIntervalError extends Error {}
